modifying terms used for human/machine access and also allowing to process without crunching logs first.

+++ main.py
@@ -7,19 +7,21 @@ import os
 import glob
 # import ipdb; ipdb.set_trace()
 
-if os.path.isfile(config.processing_database):
-    os.remove(config.processing_database)
-DbActions.create_db()
+if not config.only_calculate == True:
+    if os.path.isfile(config.processing_database):
+        os.remove(config.processing_database)
+    DbActions.create_db()
 
 print(f'Running report for {config.start_time.isoformat()} to {config.end_time.isoformat()}')
 
 # process the log lines into a sqlite database
-for lf in sorted(glob.glob(config.log_glob)):
-    with open(lf) as infile:
-        print(f'processing {lf}')
-        for line in infile:
-            ll = ip.LogLine(line)
-            ll.populate()
+if not config.only_calculate == True:
+    for lf in sorted(glob.glob(config.log_glob)):
+        with open(lf) as infile:
+            print(f'processing {lf}')
+            for line in infile:
+                ll = ip.LogLine(line)
+                ll.populate()
 
 # output for each unique identifier (that isn't robots)
 my_report = op.TsvReport()
